+++ R/lpcde/R/lpcde.R
@@ -92,7 +92,7 @@
 #' @export
 lpcde = function(x_data, y_data, y_grid=NULL, x=NULL, bw=NULL, p=NULL, q=NULL,
                  p_RBC=NULL, q_RBC=NULL, mu=NULL, nu=NULL, rbc = TRUE, ng=NULL,
-                 normalize=FALSE, nonneg=FALSE, grid_spacing=NULL,
+                 normalize=FALSE, nonneg=FALSE, grid_spacing="",
                  kernel_type=c("epanechnikov", "triangular", "uniform"),
                  bw_type=NULL){