+++ R/lpcde/NEWS.md
@@ -1,5 +1,6 @@
 # lpcde 0.1.2
 - Added flags for returning normalized and/or non-negative estimate
+- Added option to change automatic grid spacing (options include quantile and equally-spaced grid generation)
 
 # lpcde 0.1.1
 - Fixed RBC CI computation when using automatic MSE bandwidth selection.