+++ R/lpcde/NEWS.md
@@ -1,3 +1,12 @@
+# lpcde 0.1.2
+- Added flags for returning normalized and/or non-negative estimate
+- Added option to change automatic grid spacing (options include quantile and equally-spaced grid generation). Functionality available for lpcde and lpbwcde.
+- Added ng option for lpbwcde
+
+# lpcde 0.1.1
+- Fixed RBC CI computation when using automatic MSE bandwidth selection.
+- Added bandwidth regularization for MSE bandwidth selection.
+
 # lpcde 0.1.0
 - Added multivariate conditioning
 - Added tests